[HOTFIX] Avoid an error 500 on untranslated content.
This was throwing 500 errors on production upon launch (and this bad hotfix makes it shows a styled 404, which is less bad).

As authenticated user, the untranslated content is displayed, which I'm going to guess is what we actually also want for anonymous users. I will leave that as an exercise for the reader, though.